Finalização do KIDI/KIDH e correções das escoras

+++ projetoSciaXML/src/br/com/scia/xml/util/SciaXMLConstantes.java
@@ -8,7 +8,7 @@ public class SciaXMLConstantes {
 	public static final String AMBOS_OS_EIXOS = "AMBOS_OS_EIXOS";
 	public static final String INDEXADOR_NO = "N";
 	public static final String INDEXADOR_PECA = "B";
-	public static final Integer PRECISAO_ENVIO_COORDENADAS_XML = 100;
+	public static final Double PRECISAO_ENVIO_COORDENADAS_XML = 100.0;
 	public static final Double VALOR_ALTURA_MAX = 0.2;
 	public static final Double VALOR_ALTURA_MIN = 0.5;
 	public static final Double TAMANHO_MIN_ESCORA_A =  0.55;
